+++ example/lib/screens/charts_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_kit/components/navbar/navbar.dart';
-import 'package:flutter_kit/flutter_kit.dart';
-import 'package:flutter_kit/widgets/charts/chart.dart';
+import 'package:bolt_ui_kit/components/navbar/navbar.dart';
+import 'package:bolt_ui_kit/bolt_kit.dart';
+import 'package:bolt_ui_kit/widgets/charts/chart.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class ChartsScreen extends StatelessWidget {