+++ lib/widgets/inputs/inputs.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '../../theme/app_colors.dart';
+export 'pinput.dart';
 
 enum InputType {
   text,